@@ -251,46 +251,47 @@ fn asm_row_ui(ui: &mut egui::Ui, ins_diff: &ObjInsDiff, symbol: &ObjSymbol, conf
         ui.painter().rect_filled(ui.available_rect_before_wrap(), 0.0, ui.visuals().faint_bg_color);
     }
     let mut job = LayoutJob::default();
-    if let Some(ins) = &ins_diff.ins {
-        let base_color = match ins_diff.kind {
-            ObjInsDiffKind::None | ObjInsDiffKind::OpMismatch | ObjInsDiffKind::ArgMismatch => {
-                Color32::GRAY
-            }
-            ObjInsDiffKind::Replace => Color32::LIGHT_BLUE,
-            ObjInsDiffKind::Delete => COLOR_RED,
-            ObjInsDiffKind::Insert => Color32::GREEN,
-        };
+    let Some(ins) = &ins_diff.ins else {
+        ui.label("");
+        return;
+    };
+
+    let base_color = match ins_diff.kind {
+        ObjInsDiffKind::None | ObjInsDiffKind::OpMismatch | ObjInsDiffKind::ArgMismatch => {
+            Color32::GRAY
+        }
+        ObjInsDiffKind::Replace => Color32::LIGHT_BLUE,
+        ObjInsDiffKind::Delete => COLOR_RED,
+        ObjInsDiffKind::Insert => Color32::GREEN,
+    };
+    write_text(
+        &format!("{:<6}", format!("{:x}:", ins.address - symbol.address as u32)),
+        base_color,
+        &mut job,
+        config.code_font.clone(),
+    );
+    if let Some(branch) = &ins_diff.branch_from {
         write_text(
-            &format!("{:<6}", format!("{:x}:", ins.address - symbol.address as u32)),
-            base_color,
+            "~> ",
+            config.diff_colors[branch.branch_idx % config.diff_colors.len()],
             &mut job,
             config.code_font.clone(),
         );
-        if let Some(branch) = &ins_diff.branch_from {
-            write_text(
-                "~> ",
-                config.diff_colors[branch.branch_idx % config.diff_colors.len()],
-                &mut job,
-                config.code_font.clone(),
-            );
-        } else {
-            write_text("   ", base_color, &mut job, config.code_font.clone());
-        }
-        write_ins(ins, &ins_diff.kind, &ins_diff.arg_diff, symbol.address as u32, &mut job, config);
-        if let Some(branch) = &ins_diff.branch_to {
-            write_text(
-                " ~>",
-                config.diff_colors[branch.branch_idx % config.diff_colors.len()],
-                &mut job,
-                config.code_font.clone(),
-            );
-        }
-        ui.add(Label::new(job).sense(Sense::click()))
-            .on_hover_ui_at_pointer(|ui| ins_hover_ui(ui, ins))
-            .context_menu(|ui| ins_context_menu(ui, ins));
     } else {
-        ui.label("");
+        write_text("   ", base_color, &mut job, config.code_font.clone());
     }
+    write_ins(ins, &ins_diff.kind, &ins_diff.arg_diff, symbol.address as u32, &mut job, config);
+    if let Some(branch) = &ins_diff.branch_to {
+        write_text(
+            " ~>",
+            config.diff_colors[branch.branch_idx % config.diff_colors.len()],
+            &mut job,
+            config.code_font.clone(),
+        );
+    }
+    ui.add(Label::new(job).sense(Sense::click()))
+        .on_hover_ui_at_pointer(|ui| ins_hover_ui(ui, ins))
+        .context_menu(|ui| ins_context_menu(ui, ins));
 }
 
 fn asm_table_ui(
@@ -322,113 +323,105 @@ fn asm_table_ui(
 
 pub fn function_diff_ui(ui: &mut egui::Ui, view_state: &mut ViewState) -> bool {
     let mut rebuild = false;
-    if let (Some(result), Some(selected_symbol)) = (&view_state.build, &view_state.selected_symbol)
-    {
-        StripBuilder::new(ui)
-            .size(Size::exact(20.0))
-            .size(Size::exact(40.0))
-            .size(Size::remainder())
-            .vertical(|mut strip| {
-                strip.strip(|builder| {
-                    builder.sizes(Size::remainder(), 2).horizontal(|mut strip| {
-                        strip.cell(|ui| {
-                            ui.horizontal(|ui| {
-                                if ui.button("Back").clicked() {
-                                    view_state.current_view = View::SymbolDiff;
-                                }
-                            });
-                        });
-                        strip.cell(|ui| {
-                            ui.horizontal(|ui| {
-                                if ui.button("Build").clicked() {
-                                    rebuild = true;
-                                }
-                                ui.scope(|ui| {
-                                    ui.style_mut().override_text_style =
-                                        Some(egui::TextStyle::Monospace);
-                                    ui.style_mut().wrap = Some(false);
-                                    if view_state
-                                        .jobs
-                                        .iter()
-                                        .any(|job| job.job_type == Job::ObjDiff)
-                                    {
-                                        ui.label("Building...");
-                                    } else {
-                                        ui.label("Last built:");
-                                        let format =
-                                            format_description::parse("[hour]:[minute]:[second]")
-                                                .unwrap();
-                                        ui.label(
-                                            result
-                                                .time
-                                                .to_offset(view_state.utc_offset)
-                                                .format(&format)
-                                                .unwrap(),
-                                        );
-                                    }
-                                });
-                            });
+    let (Some(result), Some(selected_symbol)) = (&view_state.build, &view_state.selected_symbol) else {
+        return rebuild;
+    };
+    StripBuilder::new(ui)
+        .size(Size::exact(20.0))
+        .size(Size::exact(40.0))
+        .size(Size::remainder())
+        .vertical(|mut strip| {
+            strip.strip(|builder| {
+                builder.sizes(Size::remainder(), 2).horizontal(|mut strip| {
+                    strip.cell(|ui| {
+                        ui.horizontal(|ui| {
+                            if ui.button("Back").clicked() {
+                                view_state.current_view = View::SymbolDiff;
+                            }
                         });
                     });
-                });
-                strip.strip(|builder| {
-                    builder.sizes(Size::remainder(), 2).horizontal(|mut strip| {
-                        let demangled = demangle(&selected_symbol.symbol_name, &Default::default());
-                        strip.cell(|ui| {
+                    strip.cell(|ui| {
+                        ui.horizontal(|ui| {
+                            if ui.button("Build").clicked() {
+                                rebuild = true;
+                            }
                             ui.scope(|ui| {
                                 ui.style_mut().override_text_style =
                                     Some(egui::TextStyle::Monospace);
                                 ui.style_mut().wrap = Some(false);
-                                ui.colored_label(
-                                    Color32::WHITE,
-                                    demangled.as_ref().unwrap_or(&selected_symbol.symbol_name),
-                                );
-                                ui.label("Diff target:");
-                                ui.separator();
-                            });
-                        });
-                        strip.cell(|ui| {
-                            ui.scope(|ui| {
-                                ui.style_mut().override_text_style =
-                                    Some(egui::TextStyle::Monospace);
-                                ui.style_mut().wrap = Some(false);
-                                if let Some(match_percent) = result
-                                    .second_obj
-                                    .as_ref()
-                                    .and_then(|obj| find_symbol(obj, selected_symbol))
-                                    .and_then(|symbol| symbol.match_percent)
-                                {
-                                    ui.colored_label(
-                                        match_color_for_symbol(match_percent),
-                                        &format!("{match_percent:.0}%"),
+                                if view_state.jobs.iter().any(|job| job.job_type == Job::ObjDiff) {
+                                    ui.label("Building...");
+                                } else {
+                                    ui.label("Last built:");
+                                    let format =
+                                        format_description::parse("[hour]:[minute]:[second]")
+                                            .unwrap();
+                                    ui.label(
+                                        result
+                                            .time
+                                            .to_offset(view_state.utc_offset)
+                                            .format(&format)
+                                            .unwrap(),
                                     );
                                 }
-                                ui.label("Diff base:");
-                                ui.separator();
                             });
                         });
                     });
                 });
-                strip.cell(|ui| {
-                    if let (Some(left_obj), Some(right_obj)) =
-                        (&result.first_obj, &result.second_obj)
-                    {
-                        let table = TableBuilder::new(ui)
-                            .striped(false)
-                            .cell_layout(egui::Layout::left_to_right(egui::Align::Min))
-                            .column(Size::relative(0.5))
-                            .column(Size::relative(0.5))
-                            .resizable(false);
-                        asm_table_ui(
-                            table,
-                            left_obj,
-                            right_obj,
-                            selected_symbol,
-                            &view_state.view_config,
-                        );
-                    }
+            });
+            strip.strip(|builder| {
+                builder.sizes(Size::remainder(), 2).horizontal(|mut strip| {
+                    let demangled = demangle(&selected_symbol.symbol_name, &Default::default());
+                    strip.cell(|ui| {
+                        ui.scope(|ui| {
+                            ui.style_mut().override_text_style = Some(egui::TextStyle::Monospace);
+                            ui.style_mut().wrap = Some(false);
+                            ui.colored_label(
+                                Color32::WHITE,
+                                demangled.as_ref().unwrap_or(&selected_symbol.symbol_name),
+                            );
+                            ui.label("Diff target:");
+                            ui.separator();
+                        });
+                    });
+                    strip.cell(|ui| {
+                        ui.scope(|ui| {
+                            ui.style_mut().override_text_style = Some(egui::TextStyle::Monospace);
+                            ui.style_mut().wrap = Some(false);
+                            if let Some(match_percent) = result
+                                .second_obj
+                                .as_ref()
+                                .and_then(|obj| find_symbol(obj, selected_symbol))
+                                .and_then(|symbol| symbol.match_percent)
+                            {
+                                ui.colored_label(
+                                    match_color_for_symbol(match_percent),
+                                    &format!("{match_percent:.0}%"),
+                                );
+                            }
+                            ui.label("Diff base:");
+                            ui.separator();
+                        });
+                    });
                 });
             });
-    }
+            strip.cell(|ui| {
+                if let (Some(left_obj), Some(right_obj)) = (&result.first_obj, &result.second_obj) {
+                    let table = TableBuilder::new(ui)
+                        .striped(false)
+                        .cell_layout(egui::Layout::left_to_right(egui::Align::Min))
+                        .column(Size::relative(0.5))
+                        .column(Size::relative(0.5))
+                        .resizable(false);
+                    asm_table_ui(
+                        table,
+                        left_obj,
+                        right_obj,
+                        selected_symbol,
+                        &view_state.view_config,
+                    );
+                }
+            });
+        });
     rebuild
 }
